refactor(pathfinder): split V2Pathfinder into partial-class files (no behavior change) (#434)
V2Pathfinder.cs grew to 1834 LoC, forcing reviewers and review agents to load
the entire file plus several siblings to reason about any single change. Split
into 6 partials by pipeline stage / helper group:

- V2Pathfinder.cs              247 LoC — class shell, ctor, ComputeMaxFlow,
                                         ComputeMaxFlowWithPath, PipelineContext,
                                         ResolveAndGuard, IsBalanceNode/IsPoolNode
- V2Pathfinder.Pipeline.cs     194 LoC — stages 2–10 orchestration
- V2Pathfinder.Response.cs     252 LoC — BuildTransferDtos, BuildResponse,
                                         debug-step converters
- V2Pathfinder.Collapse.cs     372 LoC — CollapseBalanceNodes, consent filters,
                                         InsertRouterInTransfers
- V2Pathfinder.PathSteps.cs    212 LoC — MaxTransfers budget pruning + collapse
                                         helpers
- V2Pathfinder.MintOrdering.cs 184 LoC — SortEdgesForMintDependencies +
                                         invariant validator
- V2Pathfinder.Quantization.cs 437 LoC — invitation-quanta scaling + flow
                                         propagation

Pure mechanical split — no signature, body, or call-site changes. Partial
classes share the V2Pathfinder type identity so all intra-class calls remain
unqualified.

Verification (pre-push):
- Build (Circles.Pathfinder): 0 errors / 16 warnings — exact baseline match
- Build (Circles.Pathfinder.Host): 0 errors
- Tests: 1101 pass / 0 fail / 270 skip / 1371 total — exact baseline match
- Public/internal/private signature surface: 58 = 58, zero added, zero removed
- Method-body equivalence: 33/33 named method bodies byte-identical after
  whitespace normalization (SHA256). Two trivial expression-bodied one-liners
  (Sanitize, IsBalanceNode) verified manually identical.
